use u64 everywhere for dimensiosn

+++ zyx-nn/src/batch_norm.rs
@@ -30,7 +30,7 @@ pub struct BatchNorm {
 
 impl BatchNorm {
     /// Initilize layer_norm layer in device self
-    pub fn new(self, num_features: usize, dtype: DType) -> BatchNorm {
+    pub fn new(self, num_features: u64, dtype: DType) -> BatchNorm {
         BatchNorm {
             eps: 1e-5,
             momentum: 0.1,
