+++ zyx-nn/src/linear.rs
@@ -16,8 +16,8 @@ pub struct Linear {
 impl Linear {
     /// Initilize linear layer in device self
     pub fn new(
-        in_features: usize,
-        out_features: usize,
+        in_features: u64,
+        out_features: u64,
         bias: bool,
         dtype: DType,
     ) -> Result<Linear, ZyxError> {
