+++ zyx-nn/src/lstm.rs
@@ -41,7 +41,7 @@ pub struct LSTMCell {
     w_hh: Tensor,         // (4 * hidden_size, hidden_size)
     b_ih: Option<Tensor>, // (4 * hidden_size)
     b_hh: Option<Tensor>, // (4 * hidden_size)
-    hidden_size: usize,
+    hidden_size: u64,
 }
 
 impl LSTMCell {
@@ -56,8 +56,8 @@ impl LSTMCell {
     /// # Returns
     /// A `Result` wrapping the created `LSTMCell` or a `ZyxError` if initialization fails.
     pub fn new(
-        input_size: usize,
-        hidden_size: usize,
+        input_size: u64,
+        hidden_size: u64,
         bias: bool,
         dtype: Option<DType>,
     ) -> Result<Self, ZyxError> {
