+++ zyx-nn/src/rms_norm.rs
@@ -15,7 +15,7 @@ pub struct RMSNorm {
 
 impl RMSNorm {
     /// Initialize RMSNorm layer
-    pub fn new(dim: usize, dtype: DType) -> RMSNorm {
+    pub fn new(dim: u64, dtype: DType) -> RMSNorm {
         RMSNorm {
             scale: Tensor::ones(dim, dtype),
             eps: 1e-6,
