+++ zyx-nn/src/transformer_encoder_layer.rs
@@ -52,9 +52,9 @@ impl TransformerEncoderLayer {
     ///
     /// A `Result` containing the initialized `TransformerEncoderLayer` or a `ZyxError`.
     pub fn new(
-        d_model: usize,
-        nhead: usize,
-        dim_feedforward: usize,
+        d_model: u64,
+        nhead: u64,
+        dim_feedforward: u64,
         dropout: f32,
         activation: fn(Tensor) -> Tensor,
         layer_norm_eps: f64,
